+++ src/main/java/fr/julien/packagedautoterminals/network/PatNetwork.java
@@ -19,5 +19,6 @@ public final class PatNetwork {
         CHANNEL.registerMessage(PacketEditorFill.Handler.class, PacketEditorFill.class, 3, Side.SERVER);
         CHANNEL.registerMessage(PacketRenameGroup.Handler.class, PacketRenameGroup.class, 4, Side.SERVER);
         CHANNEL.registerMessage(PacketOpenTerminal.Handler.class, PacketOpenTerminal.class, 5, Side.SERVER);
+        CHANNEL.registerMessage(PacketEditorGhost.Handler.class, PacketEditorGhost.class, 6, Side.SERVER);
     }
 }